Insert synchronization ops after asycn copy creation. (#38)
Summary:
With arbitary data channel the producer ops could be a non-load op,
which mean it doesn't necessarily write to SMEM. A `local_store` op will
be created to support the data channel, and in that case the producer
acquire primitive should precede the `local_store` op instead of the
original producer op. This change enables that by creating the async
copy before inserting synchronization primitives. A map is set up to
track the created async copy and the original data channel.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: