Update bufferIdx correctly to handle IfOp under ForOp (#55)
Summary:
For the case of
```
ForOp
  IfOp
    ForOp
      Channel
```
We need to create AccumLoopCount for the channel through the outer
ForOp. Prior to the fix, bufIdx argument of the outer loop increases by
one for each iteration, which can cause deadlock when the IfOp is not
executed for some iterations.

The fix is to add AccumIf argument for the outer ForOp:
```
ForOp accumIf, bufIdx
  IfOp
    ForOp
      Channel uses (idx, phase) based on accumIf
    accumIf += 1
```
Modify createNewLoop to handle the general case instead of depending on
    bool hasParallelReuse, bool isOuterOfReuse.